+++ packages/jael/angel_jael/lib/angel3_jael.dart
@@ -17,11 +17,12 @@ AngelConfigurer jael(Directory viewsDirectory,
     {String fileExtension = '.jael',
     bool strictResolution = false,
     bool cacheViews = true,
+    Map<String, Document>? cache,
     Iterable<Patcher> patch = const [],
     bool asDSX = false,
     bool minified = true,
     CodeBuffer Function()? createBuffer}) {
-  var cache = <String, Document>{};
+  var _cache = cache ?? <String, Document>{};
 
   var bufferFunc = createBuffer ?? () => CodeBuffer();
 
@@ -34,31 +35,20 @@ AngelConfigurer jael(Directory viewsDirectory,
       var errors = <JaelError>[];
       Document? processed;
 
-      if (cacheViews && cache.containsKey(name)) {
-        processed = cache[name];
-      } else {
-        var file = viewsDirectory.childFile(name + fileExtension);
-        var contents = await file.readAsString();
-        var doc = parseDocument(contents,
-            sourceUrl: file.uri, asDSX: asDSX, onError: errors.add);
-        if (doc == null) {
-          throw ArgumentError(name + fileExtension + " does not exists");
-        }
+      //var stopwatch = Stopwatch()..start();
 
-        try {
-          processed = await (resolve(doc, viewsDirectory,
-              patch: patch, onError: errors.add));
-        } catch (e) {
-          // Ignore these errors, so that we can show syntax errors.
-        }
-        if (processed == null) {
-          throw ArgumentError(name + fileExtension + " does not exists");
-        }
+      if (cacheViews && _cache.containsKey(name)) {
+        processed = _cache[name];
+      } else {
+        processed = await _loadViewTemplate(viewsDirectory, name,
+            fileExtension: fileExtension, asDSX: asDSX, patch: patch);
 
         if (cacheViews) {
-          cache[name] = processed;
+          _cache[name] = processed!;
         }
       }
+      //print('Time executed: ${stopwatch.elapsed.inMilliseconds}');
+      //stopwatch.stop();
 
       var buf = bufferFunc();
       var scope = SymbolTable(
@@ -69,7 +59,7 @@ AngelConfigurer jael(Directory viewsDirectory,
       if (errors.isEmpty) {
         try {
           const Renderer().render(processed!, buf, scope,
-              strictResolution: strictResolution == true);
+              strictResolution: strictResolution);
           return buf.toString();
         } on JaelError catch (e) {
           errors.add(e);
@@ -80,4 +70,55 @@ AngelConfigurer jael(Directory viewsDirectory,
       return buf.toString();
     };
   };
+}
+
+/// Preload all of Jael templates into a cache
+///
+///
+/// To apply additional transforms to parsed documents, provide a set of [patch] functions.
+Future<void> jaelTemplatePreload(
+    Directory viewsDirectory, Map<String, Document> cache,
+    {String fileExtension = '.jael',
+    bool asDSX = false,
+    Iterable<Patcher> patch = const []}) async {
+  await viewsDirectory.list(recursive: true).forEach((f) async {
+    if (f.basename.endsWith(fileExtension)) {
+      var name = f.basename.split(".");
+      if (name.length > 1) {
+        //print("View: ${name[0]}");
+        Document? processed = await _loadViewTemplate(viewsDirectory, name[0]);
+        if (processed != null) {
+          cache[name[0]] = processed;
+        }
+      }
+    }
+  });
+}
+
+Future<Document?> _loadViewTemplate(Directory viewsDirectory, String name,
+    {String fileExtension = '.jael',
+    bool asDSX = false,
+    Iterable<Patcher> patch = const []}) async {
+  var errors = <JaelError>[];
+  Document? processed;
+
+  var file = viewsDirectory.childFile(name + fileExtension);
+  var contents = await file.readAsString();
+  var doc = parseDocument(contents,
+      sourceUrl: file.uri, asDSX: asDSX, onError: errors.add);
+
+  if (doc == null) {
+    throw ArgumentError(file.basename + " does not exists");
+  }
+
+  try {
+    processed =
+        await (resolve(doc, viewsDirectory, patch: patch, onError: errors.add));
+  } catch (e) {
+    // Ignore these errors, so that we can show syntax errors.
+  }
+  if (processed == null) {
+    throw ArgumentError(file.basename + " does not exists");
+  }
+  return processed;
 }